Lab_1/lab1_analysis/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
psycopg2-binary>=2.9.5
sqlalchemy>=1.4.45
openpyxl>=3.0.0